feature: Getting `param_grid`.
